Educational_AI/apps/api/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Firebase
firebase-admin>=6.3.0
google-cloud-firestore>=2.14.0
google-cloud-storage>=2.14.0

# LLM Providers
# TODO: [Phase 1] Uncomment when implementing real providers
google-cloud-aiplatform>=1.38.0
# openai>=1.6.0

# Document Parsing
pypdf>=6.0.0
# unstructured[pdf,docx,pptx]>=0.11.0
# python-pptx>=0.6.21
# python-docx>=1.1.0

# Vector Search (v1: in-memory numpy)
numpy>=1.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.25.0  # For TestClient
